xanhsm_demo/xanhsm/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-core
langchain-openai
langgraph
openai
